break down team creation and athlete profile into smaller files
