+++ src/components/player/profile-modal-components/tabs/OverviewTab.tsx
@@ -1,0 +1,50 @@
+import React from 'react';
+import { formatDate } from '../utils';
+
+interface OverviewTabProps {
+  player: any;
+}
+
+export const OverviewTab: React.FC<OverviewTabProps> = ({ player }) => {
+  return (
+    <div className="space-y-4 p-4">
+      <div className="grid grid-cols-2 gap-4">
+        <div className="text-gray-500 dark:text-gray-400">Name</div>
+        <div className="font-medium text-gray-700 dark:text-gray-300">{player.player_name}</div>
+        
+        <div className="text-gray-500 dark:text-gray-400">Team</div>
+        <div className="font-medium text-gray-700 dark:text-gray-300">{player.team_name}</div>
+        
+        <div className="text-gray-500 dark:text-gray-400">Date of Birth</div>
+        <div className="font-medium text-gray-700 dark:text-gray-300">{formatDate(player.date_of_birth)}</div>
+        
+        {player.height && (
+          <>
+            <div className="text-gray-500 dark:text-gray-400">Height</div>
+            <div className="font-medium text-gray-700 dark:text-gray-300">
+              {player.height} cm ({Math.floor(player.height / 30.48)}'{Math.round((player.height / 2.54) % 12)}")
+            </div>
+          </>
+        )}
+        
+        {player.weight && (
+          <>
+            <div className="text-gray-500 dark:text-gray-400">Weight</div>
+            <div className="font-medium text-gray-700 dark:text-gray-300">
+              {player.weight} kg ({Math.round(player.weight * 2.20462)} lbs)
+            </div>
+          </>
+        )}
+        
+        {player.birth_country && (
+          <>
+            <div className="text-gray-500 dark:text-gray-400">Country of Birth</div>
+            <div className="font-medium text-gray-700 dark:text-gray-300">{player.birth_country}</div>
+          </>
+        )}
+      </div>
+    </div>
+  );
+};
+
+export default OverviewTab;
