+++ src/components/team-creation/player-selection-components/SearchBar.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+
+interface SearchBarProps {
+  searchQuery: string;
+  onSearchChange: (query: string) => void;
+}
+
+export const SearchBar: React.FC<SearchBarProps> = ({ searchQuery, onSearchChange }) => {
+  return (
+    <div className="px-6 py-4 border-b dark:border-gray-700">
+      <div className="relative">
+        <span className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400">🔍</span>
+        <input
+          type="text"
+          placeholder="Search players or teams..."
+          value={searchQuery}
+          onChange={(e) => onSearchChange(e.target.value)}
+          className="w-full pl-10 pr-4 py-2 border dark:border-gray-600 rounded-lg focus:outline-none focus:ring-2 focus:ring-green-500 focus:border-transparent dark:bg-dark-700 dark:text-gray-100"
+        />
+      </div>
+    </div>
+  );
+};
+
+export default SearchBar;
